requirements.txt
pandas>=1.3.0
firebase-admin>=5.0.0
openpyxl>=3.0.0
xlrd>=2.0.1
watchdog>=2.1.0
python-dotenv>=0.19.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.12.0